MNIST_Model/Reference/ImageNet_ResNet50/requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.0.0

# Data Augmentation
albumentations>=1.3.0
opencv-python-headless>=4.8.0

# Visualization & Metrics
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Grad-CAM Analysis
pytorch-grad-cam>=1.4.0
grad-cam>=1.4.0

# Progress Bars
tqdm>=4.65.0

# Model Summary
torchsummary>=1.5.1

# Deployment (for Phase 4)
gradio>=4.0.0
huggingface-hub>=0.19.0

# Optional: Experiment Tracking
# tensorboard>=2.13.0
# wandb>=0.15.0

# Jupyter Support (for notebooks)
jupyter>=1.0.0
ipywidgets>=8.0.0
nbformat>=5.9.0

